feat: Support Project Name generation
As when defining different workspaces for each directory,
this info is also needed when running `atlantis plan/apply -d ${git_root}/stage/app -w stage_app`
to run the command on specific directory but one could also use the
`atlantis plan/apply -p stage_app` in case there is a project name defined.
As this uses the same convention with the name of workspace (which should
be unique) and the default value is disabled this should not create any
problems to the current users.