Support & Enable Cross Compilation of the Project for multiple Platforms & Other Optimizations (#58)
* improvement: Add Phony Targets in Makefile

As described in better detail in this stackoverflow https://stackoverflow.com/questions/2145590/what-is-the-purpose-of-phony-in-a-makefile
the phony targets do not represent file targets and this is what we want
in our case.

* improvement: Support Cross-Compilation

Changes required to support cross-compilation for multi platforms.
This keeps all the old Make Targets & fixes the ones needed, so no side-
effects are expected.

* feat: Enable Cross-compilation for multi-Platforms

Enable the cross-compilation for multiple platforms and generate their
artifacts & binaries. This will allow to push artifacts for each new version
in Github Releases, among with SHA256SUMS and users will be able to download
those binaries from all platforms.

* improvement: Drop Hardcoded Version from code

With this change, version is only defined in one place, Makefile as a
hardcoded text. However this may change in the future and this change will
enable the version value to be managed easier & reduce mistakes.