Migrate to Gox for cross-compilation
This enables us to build for platforms such as Apple Silicon,
which was the main motivation for the change.
Goxc doesn't support this:

```
[goxc] 2023/02/09 01:34:10 WARNING: Architecture 'arm64' is unsupported
```

and its repository has been in read-only mode since June 2022.

Since Gox is a much simpler project, it doesn't support packaging zip
and tar archives. Implement this ourselves, while maintaining archive
internal paths.