Fix parsing environment variables
This was previously breaking tests for me locally, as well as on the CI.

It appears to me that if we were to use `SplitN`, we would have to
specify the `n = 2` to get at most 2 substrings: the name and the value.
However, the documentation around `SplitN` recommends we use `Cut`
instead, which seems simpler.